+++ lib/features/number_trivia/data/model/number_trivia_model.dart
@@ -2,12 +2,12 @@ import 'package:meta/meta.dart';
 import 'package:learning/features/number_trivia/domain/entities/number_trivia.dart';
 
 class NumberTriviaModel extends NumberTrivia {
-  NumberTriviaModel({@required String text, @required double number})
+  NumberTriviaModel({@required String text, @required int number})
       : super(text: text, number: number);
 
   factory NumberTriviaModel.fromJSON(Map<String, dynamic> json) {
     return NumberTriviaModel(
-        text: json['text'], number: (json['number'] as num).toDouble());
+        text: json['text'], number: (json['number'] as num).toInt());
   }
 
   Map<String, dynamic> toJSON() {
